Nav, Header and Footer work
Pushed the status indicator to the right of the navbar and put the light to the right of the text.
Seperated the header from the main content on every page and gave buttons within those headers css and created a seperator.
Gave the footer a fixed size, some css and stuck it to the bottom of the page.